Overall refactoring of OutflowProcess constructor to improve accuracy and stability. Corrected an error which allowed unchoked tail of outflow process to be ignored. Boundary conditions now implemented without fictitious points. Quadratic Simpson integration rather than trapezoidal. Improved internal documentation. All functionality tested; no major changes in results when used in aRCPlan.

+++ Source/File.h
@@ -35,7 +35,7 @@ private:
     string temp_dir;
     string subfolder;
     string file_name;
-    int found;
+    unsigned long found;
     int log_number;
 
 public:
